requirements.txt
ib_async>=1.0.0
aiohttp>=3.9.0
pydantic>=2.5.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
